credit_policy_converter/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
anthropic>=0.30.0
openpyxl>=3.1.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
pydantic>=2.6.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
